requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.0.0
python-dotenv>=1.0.0
google-genai>=1.0.0
replicate>=1.0.0
edge-tts>=6.1.0
mutagen>=1.47.0
httpx>=0.27.0
aiofiles>=23.0.0
jinja2>=3.1.0